+++ src/AcountPopups/RegisterAccount.js
@@ -1,25 +1,26 @@
-import React, {useContext, useState} from "react";
-import RegisterAccountPopup from "AcountPopups/CreateNewAccountPopup";
-import AccountLoginPopup from "./AccountLoginPopup";
+import React, { useContext, useState } from "react";
+import RegisterAccountPopup from "./RegisterAccountPopup.js";
 
-function RegisterAccount()
-{
-    const [registrationPopupActive, setRegistrationPopupActive] = useState(false);
-    return (
-        <>
-        <button onClick={() => {
+function RegisterAccount() {
+  const [registrationPopupActive, setRegistrationPopupActive] = useState(false);
 
-        setRegistrationPopupActive(true);
-    }}
-                    className="create-account-btn blue"
-    >Sukurti naują paskyrą
-    </button>
-            {registrationPopupActive && (
-    <RegisterAccountPopup
-        popupActive={registrationPopupActive}
-        setpopupActive={setRegistrationPopupActive}
-    />)}
-        </>
-    );
+  return (
+    <>
+      <button
+        onClick={() => {
+          setRegistrationPopupActive(true);
+        }}
+        className="create-account-btn blue"
+      >
+        Sukurti naują paskyrą
+      </button>
+      {registrationPopupActive && (
+        <RegisterAccountPopup
+          popupActive={registrationPopupActive}
+          setpopupActive={setRegistrationPopupActive}
+        />
+      )}
+    </>
+  );
 }
 export default RegisterAccount;
