+++ src/App.js
@@ -6,7 +6,7 @@ import Header from "Header/Header";
 import ShoppingCartPage from "./ShoppingCartPage/ShoppingCartPage";
 import { useState } from "react";
 import { GlobalProvider } from "Context/GlobalState";
-
+import RegisterAccount from "./AcountPopups/RegisterAccount";
 
 function App() {
   const [headerTitle, setHeaderTitle] = useState("Gėlių e-parduotuvė");
@@ -18,7 +18,6 @@ function App() {
       <Router>
         <div className="App">
           <Header title={headerTitle} />
-
           <Route
             exact
             path="/"
